Allow for labeled metric queries
This commit allows for querying for information (both from the model and
from the Oldtimer) about labeled metrics.  This is done by passing the
`labels` query parameter when querying for metric data.  The value of
the `labels` parameter should be a comma-separated list of `key:value` pairs.
The label set provided must be an exact match for the set of "extra" labels
on the target metric.